chore: Merge parsers/ into poland spider
